+++ runtime/docs/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add a new MinBlockTimeMsec config field to consensus
+
 ## [0.0.0.44] - 2023-06-26
 
 - Add a new ServiceConfig field to servicer config